Fix sp-kill-hybrid-sexp bug in REPL-like modes
After killing a full line (i.e. sp-point-in-blank-line returns non-nil),
sp-kill-hybrid-sexp inserts a number of spaces equal to the value
of (current-column) before the kill was performed. When killing a full
line in a mode where the editable line begins at a non-zero column this
causes extraneous spaces to be inserted.

Instead, only insert the spaces (if any) necessary to get point back to
the original indentation level recorded before the kill.