[Fix #495] Fix sp--next-thing-selection at the end of sexp.
If the point is after the last sexp inside the list and before its
closing delimiter, we should return just the whitespace.  This is
consistent with sp-kill-sexp and also with the philosophy that we should
operate on everything left *in* the list, which is nothing.  Note that
selecting next sexp with argument 1 will select the parent as usual.