Don't indent if aggressive-indent-mode is enabled
aggressive-indent-mode [1] already provides support for re-indentation
after buffer changes, and it does it more sophisticated than just
calling indent-region (and additionally, it won't issue an annoying
"Indenting region..." message).  Therefore, if aggressive-indent-mode is
enabled in the current buffer, don't indent the region but let the job
be done by that.

[1] https://github.com/Malabarba/aggressive-indent-mode