Don't ignore errors when running :when and :unless functions.
Otherwise, this can lead to very hard-to-debug errors.

Consider the code:

    (defun always-true (id context)
      t)

    (sp-with-modes '(some-mode)
      (sp-local-pair "[" "]" :unless '(always-true)))

We would expect [ to never be automatically paired. However,
`always-true` only takes 2 arguments, when it should take 3. As a
result, we swallow the error and consider `always-true` to have returned
nil!

This made my debugging tricky, so I believe throwing the error
sooner is the right thing to do here.