Interactive behaviour for sp-{next,previous}-sexp
The current behaviour for sp-{next,previous} is unintiutive in
interactive mode when the point is not on the :end or :beg respectively.
Users may expect the following behaviour:

   (f|oo bar) -> (foo |bar) ;; sp-next-sexp

   (foo b|ar baz) -> (foo| bar baz) ;; sp-previous-sexp

This commit introduces a new defcustom, `sp-interactive-dwim' (nil by
default) to achieve the desired behaviour. It is only activated if
sp-{next,previous}-sexp is called interactively.

Note: the (if ...) statements of both implementations have been reversed
in order to avoid introducing a (progn ...) for the true case.