Removed pairing for backtick and apostrophe.
It seems that sp-latex-skip-match-apostrophe doesn't perform its stated
function, because placing the cursor before a quote (say, in math mode,
where quote is used to denote a derivative) will result in "Unmatched
expression." showing up in the minibuffer.

Note that smartparens, in an attempt to find the matching backtick for
the single quote, substantially slows down editing in any (La)TeX file
containing unmatched quotes (e.g. derivative f'(x) showing up in math mode).

Since single quotes appear relatively infrequently in English prose, the
loss in functionality is minimal and the speed-up benefit seems to be
worth it.