Fix autoskipping closing chevrons in rust-mode (#651)
Before this commit, closing chevrons were not autoskipped.  Typing:

`T<` would result in the correct insertion of a matching closing
chevron: `T<|>`, but typing `>` would not result in `T<>|>`, not
skipping the closing chevron even with the `autoskip` action set on the
pair.

This was caused by the regexp in `sp-rust-could-be-parameterized` set as
a `when` filter for the local pair "<" ">".  This filter is called by
`sp--do-action-p` not only when inserting a chevron, but also when
determining if the closing chevron should be autoskipped.  In the
closing case, the regexp does not match, because at the moment this
check is done the code contains one extra chevron: `<T>|>`.

In fact, the filter `sp-rust-could-be-parameterized` was implicitly
concerned only with the `insert` action, and not the other ones.

This commits rewrites the filter to explicitly disallow some cases where
a matching bracket should not be inserted.

Closes #642.