Do not retry parsing after first failed sexp.
The previous state was likely a hack to solve some problem.  But if we
ask for next sexp and it is broken, we should respect that.  If "next
valid" is really required, we can wrap the call in a while loop or
something like that and jump after the invalid delimiters.

Also make sure to fail properly if there are no stringlike or no regular
pairs defined (the regexps return "" which match anything and that gives
wrong results)