Don't affect delete-selection-mode when SP is off
When delete-selection-mode is on, typing a character with an active
region will replace the region with the typed character.  However,
smartparens advises delete-selection-pre-hook to inhibit this behavior
when the typed character is part of a pair, as smartparens wants to
wrap the region rather than replace it.

However, if you turn smartparens on and then back off in a buffer
where delete-selection-mode is also on, this advice would prevent
delete-selection-pre-hook from replacing the selection even though
smartparens is not on, and therefore the region would be neither
deleted nor wrapped.  This commit makes it so that the smartparens
advice never affects delete-selection-pre-hook unless smartparens-mode
is turned on.

Fixes #763.