sp-kill-whole-line creates single kill-ring enrty
sp-kill-whole-line now creates only on (bigger) kill-ring entry: the killed text and the killed new-line. This makes its behaviour more consistent with the built-in commands  'kill-line and 'kill-whole-line, when it comes to yanking. Also repeated invocations or combinations with other kill-commands will behave this way, which is more compliant with the emacs-conventions.