Also detect minibuffer-mode
Recently, on Emacs master branch, a commit 55db25b257 was pushed to fix
Emacs bug 47150. This bug was related to an incorrect implementation of
minibuffer-inactive-mode. It used to be the case that after the first
usage of the minibuffer, it stucks in minibuffer-inactive-mode, though
minibuffer-inactive-mode was intended for a completely different
purpose, and the minibuffer on the current frame should NOT be inactive.
The new design is to have two possible major modes for the minibuffer:
minibuffer-mode when it is active, and minibuffer-inactive-mode when it
is inactive.