fix(syntax): use syntax-after instead char-syntax to determine buffer syntax
Function char-syntax does not take into account the syntax-table text
property which might be present in the buffer.  Sometimes one
character is used in multiple syntax contexts and Emacs solves this by
using text properties on specific places to hint what is the true
syntax.  This was causing subtle bugs where we were using the default
syntax info instead of the concrete one placed in the buffer.

Fixes #851, #1195.