feat: rework how globalized modes are initialized
The sp-turn-off-* functions are marked as obsolete because they serve
no purpose, every user can implement them with a lambda if necessary.

The smartparens mode and strict mode now use common code to turn
themselves on.

The show-smartparens-mode only uses sp-ignore-modes-list and turns
itself on in comint and special modes automatically.